requirements-anomaly.txt
# Deep learning
torch>=2.0
lightning>=2.0

# Config management
hydra-core>=1.3
omegaconf>=2.3

# ML & data
scikit-learn>=1.3
pandas
numpy
scipy

# Experiment tracking
wandb

# Already in requirements.txt (duplicated here for self-contained install)
duckdb
pyyaml
ruptures
